requirements.txt
flask
requests
kafka-python
snowflake-connector-python
watchdog

